#53 Bug fix. Buttons are now stored in the DB properly.

+++ Android/app/src/main/java/com/indra/indra/models/RemoteButtonModel.java
@@ -35,7 +35,7 @@ public class RemoteButtonModel {
         return remoteId;
     }
 
-    public void setRemoteId(int remoteId) {
+    public void setRemoteId(long remoteId) {
         this.remoteId = remoteId;
     }
